move refresh token function back into helpers file; only calling function from new file run by Scheduler.
