+++ server/utils/helper.js
@@ -234,7 +234,20 @@ var getUserData = function(userId, req, res) {
     });
 };
 
+// decrypt refresh token, renew access token, and update data in database
+var refreshUserAccessToken = function() {
+  db.Users.findAll()
+    .then(function (users) {
+      var apiInterval = 0;
+      users.forEach(function(user, key, collection) {
+        var decipher = crypto.createDecipher(process.env.CIPHER_ALGORITHM, process.env.CIPHER_KEY);
+        var decryptedRefreshToken = decipher.update(user.refreshToken, 'hex', 'utf8') + decipher.final('utf8');
+        setTimeout(function() {sliceRefreshRequest(decryptedRefreshToken, saveUpdatedTokens, user.id);}, apiInterval);
+        apiInterval += 15000; // space the API request for each user out by 15 seconds
+      });
+    });
+};
+
 module.exports.getUserData = getUserData;
-module.exports.saveUpdatedTokens = saveUpdatedTokens;
-module.exports.sliceRefreshRequest = sliceRefreshRequest;
+module.exports.refreshUserAccessToken = refreshUserAccessToken;
 module.exports.ensureAuthenticated = ensureAuthenticated;
